Backend/House_Rent/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==3.2.20
django-adminlte-3==0.1.6
django-adminlte2==0.4.3
django-cors-headers==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
django-material-admin==1.8.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
jsonschema==4.17.3
Markdown==3.4.4
pkgutil_resolve_name==1.3.10
PyJWT==2.8.0
pyproject-toml==0.0.10
pyrsistent==0.19.3
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
zipp==3.15.0
